完成部分Service层逻辑

+++ server-springboot/src/main/java/com/gitee/swsk33/gitdocument/dao/AnthologyDAO.java
@@ -10,11 +10,11 @@ public interface AnthologyDAO {
 
 	int add(Anthology anthology);
 
-	int delete(int id);
+	int delete(long id);
 
 	int update(Anthology anthology);
 
-	Anthology getById(int id);
+	Anthology getById(long id);
 
 	List<Anthology> getAll();
